+++ NEWS.md
@@ -1,4 +1,4 @@
-# schtools development version
+# schtools 0.3.0
 
 - New function `log_snakemake()` redirects all output to a rule-specific log
   file when called by a Snakemake rule (#35, @kelly-sovacool).